+++ lib/src/extensions/index.dart
@@ -1,0 +1,1 @@
+export 'ttext_editing_controller_extension.dart';